license: add copyright line for all contributors
Recently it has come up that someone wants to re-use part of
wlroots in their own project. The standard procedure is to copy
over the license text and copyright lines.

Let's make it clear that the wlroots copyright is held by all
wlroots contributors, and that the license file comes from the
wlroots project.

See [1] for more details.

[1]: https://www.linuxfoundation.org/blog/blog/copyright-notices-in-open-source-software-projects