build: automatically compute soversion
Instead of requiring a manual soversion bump each time we do a
release, automatically compute it from the version number.

In the early days we used to not bump soversion, so there is an
hardcoded offset.

This will need to be updated once we ship 1.0, so assert that we're
still on 0.x.